Create README.md
I am a Fresh Graduate and my major is Electrical Engineering. I am keen on reading in Cyper Security as well as discovering new Languages as I believe it can extend my horizon to a new level and help me in integrating in a new relationships. I also love reading web-novels whether it is eastern or western as a part of my free time